+++ source/Neural Network/Internal/Dense.h
@@ -24,7 +24,7 @@ namespace OddityEngine::NeuralNetwork {
             return this->weights * input + this->bias;
         };
 
-        Vector<> backward(Vector<> output_gradient, const double learning_rate) override {
+        Vector<> backward(Vector<> output_gradient, const float learning_rate) override {
             auto weights_gradient = output_gradient * (transpose(this->input));
             this->weights -= weights_gradient * learning_rate;
             this->bias -= output_gradient * learning_rate;
